Updated ReminderForm to use more standard SwiftUI
Replaced all of ReminderForm's State(wrappedValue:) with a Binding. This made it possible to remove ReminderForm's custom init. Added ReminderFormConfig for the form's data and logic. Change from .sheet(item:) which is for static content to .sheet(isPresented:) which is for dynamic. Replaced the Environment dismiss with isPresented = false.

There are other instances of State(wrappedValue:) in the project which could also be replaced with Binding.