Compute statement hashability from underlying query fragment
We currently depend directly on the hashability of `Statement` in the
various fetch property wrappers, but this conformance has been found to
be problematic and so we should remove it. We can make this change
today, though, before removing it, by computing the hashability of a
statement from its query fragment data.