Support `attachMetadatabase` in previews (#267)
* Support `attachMetadatabase` in previews

Previews can currently crash if an app database is provisioned that
attaches a metadatabase. Despite trying to ensure all databases are
in-memory to work around an Xcode previews quirk, SQLite still throws a
file system error when connecting to an in-memory database with a file
URL.

This PR attempts to work around the issue by sharing the same database
connection in previews.

This does require losing the internal `DatabaseMigrator` we currently
use for the metadatabase and instead depend on `IF NOT EXISTS` because
GRDB only supports a single set of migrations per database (we could
maybe scope this change just to previews), but this is probably OK. It
also means that the migrator will always detect schema changes for
previews since the metadatabase tables will conflict, but this is also
probably OK since previews are not long-living.

* Revert "Support `attachMetadatabase` in previews"

This reverts commit 173471163096cae57b2c82d00499184ca23877ff.

* wip

* wip

* wip

* wip

---------

Co-authored-by: Brandon Williams <mbrandonw@hey.com>